+++ server/index.js
@@ -3,7 +3,8 @@ require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const cors = require('cors');
-const routes = require('./routes/routes');
+const eventRoutes = require('./routes/event');
+const profileRoutes = require('./routes/profile');
 
 const mongoString = process.env.DATABASE_URL;
 const mongoPassword = process.env.npm_config_db_pass;
@@ -26,7 +27,8 @@ const app = express();
 
 app.use(express.json());
 app.use(cors());
-app.use('/api', routes);
+app.use('/api/events', eventRoutes);
+app.use('/api/profiles', profileRoutes);
 
 app.listen(3000, () => {
     console.log(`Server started at ${3000}`);
